feature: errors treatment

+++ src/favorites/favorites.service.ts
@@ -1,13 +1,28 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { Prisma } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { Product } from 'src/products/entities/product.entity';
+import { User } from 'src/users/entities/users.entities';
+import { handleError } from 'src/utils/handle-error-unique.util';
 import { FavoriteProductDto } from './dto/favorite-product.dto';
 import { Favorite } from './entities/favorite.entity';
 
 @Injectable()
 export class FavoritesService {
   constructor(private readonly prisma: PrismaService) {}
-  favoriteProduct(dto: FavoriteProductDto): Promise<Favorite> {
+  async favoriteProduct(dto: FavoriteProductDto): Promise<Favorite> {
+    await this.verifyId(dto.userId);
+
+    const product: Product = await this.prisma.product.findUnique({
+      where: { name: dto.productName },
+    });
+
+    if (!product) {
+      throw new NotFoundException(
+        `Produto de nome ${dto.productName} não encontrado`,
+      );
+    }
+
     const data: Prisma.FavoriteCreateInput = {
       user: {
         connect: {
@@ -21,18 +36,45 @@ export class FavoritesService {
       },
     };
 
-    return this.prisma.favorite.create({ data });
+    return this.prisma.favorite.create({ data }).catch(handleError);
   }
 
-  unfavoriteProduct(id: string) {
+  async unfavoriteProduct(id: string) {
+    const favorite: Favorite = await this.prisma.favorite.findUnique({
+      where: { id },
+    });
+
+    if (!favorite) {
+      throw new NotFoundException(`Entrada de id ${id} nao encontrada`);
+    }
+
     return this.prisma.favorite.delete({ where: { id } });
   }
 
-  getUserFavorites(id: string): Promise<Favorite[]> {
+  async getUserFavorites(id: string): Promise<Favorite[]> {
+    await this.verifyId(id);
     return this.prisma.favorite.findMany({ where: { userId: id } });
   }
 
-  getUsersWhoFavoritedProduct(id: string) {
+  async getUsersWhoFavoritedProduct(id: string) {
+    const product: Product = await this.prisma.product.findUnique({
+      where: { id },
+    });
+
+    if (!product) {
+      throw new NotFoundException(`Entrada de id ${id} nao encontrada`);
+    }
+
     return this.prisma.favorite.findMany({ where: { product: { id } } });
+  }
+
+  async verifyId(id: string): Promise<void | never> {
+    const user: User = await this.prisma.user.findUnique({
+      where: { id },
+    });
+
+    if (!user) {
+      throw new NotFoundException(`Entrada de id ${id} nao encontrada`);
+    }
   }
 }
